+++ assets/javascripts/kitten/components/tables/double-entry-table/index.js
@@ -1,81 +1,47 @@
-import React, { Component } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import { Header } from './components/header'
-import { Col, HeaderCol, TitleCol } from './components/col'
-import styled from 'styled-components'
-import COLORS from '../../../constants/colors-config'
-import { pxToRem } from '../../../helpers/utils/typography'
-import { hexToRgba as rgba } from '../../../helpers/utils/hex-to-rgba'
-import { ScreenConfig } from '../../../constants/screen-config'
+import { Styles } from './styles'
 
-const StyledContainer = styled.div`
-  position: relative;
-`
-
-const StyledTableContainer = styled.div`
-  overflow-x: scroll;
-  margin-bottom: 1.875rem;
-  border-left: ${pxToRem(2)} solid ${rgba(COLORS.background1, 0.0667)};
-  border-right: ${pxToRem(2)} solid ${rgba(COLORS.background1, 0.0667)};
-
-  background: linear-gradient(
-      to right,
-      white 30%,
-      ${rgba(COLORS.background1, 0)}
-    ),
-    linear-gradient(to left, white 30%, ${rgba(COLORS.background1, 0)}),
-    radial-gradient(
-      farthest-side at 0 50%,
-      ${rgba(COLORS.font1, 0.2)},
-      ${rgba(COLORS.font1, 0)}
-    ),
-    radial-gradient(
-      farthest-side at 100% 50%,
-      ${rgba(COLORS.font1, 0.2)},
-      ${rgba(COLORS.font1, 0)}
-    );
-
-  background-repeat: no-repeat;
-  background-size: ${pxToRem(40)} 100%, ${pxToRem(40)} 100%, ${pxToRem(14)} 100%,
-    ${pxToRem(14)} 100%;
-  background-position: ${pxToRem(120)}, 100%;
-
-  /* Opera doesn't support this in the shorthand */
-  background-attachment: local, local, scroll, scroll;
-
-  @media screen and (min-width: 0\0) {
-    /* IE */
-    background: ${COLORS.background1};
-  }
-
-  @media (min-width: ${pxToRem(ScreenConfig.M.min)}) {
-    background-position: ${pxToRem(210)}, 100%;
-  }
-`
-
-const StyledTable = styled.table`
-  width: 100%;
-  border-spacing: 0;
-  table-layout: fixed;
-  border-collapse: separate;
-
-  @media screen and (min-width: 0\0) {
-    /* IE */
-    border-collapse: collapse;
-  }
-`
-
-export const DoubleEntryTable = props => (
-  <StyledContainer>
-    <StyledTableContainer>
-      <StyledTable {...props} />
-    </StyledTableContainer>
-  </StyledContainer>
+export const DoubleEntryTable = ({ firstColWidth, ...props }) => (
+  <Styles className="DoubleEntryTable__Container" firstColWidth={firstColWidth}>
+    <div className="DoubleEntryTable__TableContainer">
+      <table className="DoubleEntryTable__Table" {...props} />
+    </div>
+  </Styles>
 )
 
-DoubleEntryTable.Header = Header
+DoubleEntryTable.Header = ({ children, headerRowProps, ...others }) => (
+  <thead {...others}>
+    <tr {...headerRowProps}>{children}</tr>
+  </thead>
+)
 DoubleEntryTable.Body = props => <tbody {...props} />
 DoubleEntryTable.Row = props => <tr {...props} />
-DoubleEntryTable.Col = Col
-DoubleEntryTable.HeaderCol = HeaderCol
-DoubleEntryTable.TitleCol = TitleCol
+DoubleEntryTable.Col = props => (
+  <td
+    className="DoubleEntryTable__Column DoubleEntryTable__Column--Col"
+    {...props}
+  />
+)
+DoubleEntryTable.HeaderCol = props => (
+  <th
+    className="DoubleEntryTable__Column DoubleEntryTable__Column--HeaderCol"
+    scope="column"
+    {...props}
+  />
+)
+DoubleEntryTable.TitleCol = props => (
+  <th
+    className="DoubleEntryTable__Column DoubleEntryTable__Column--TitleCol"
+    scope="row"
+    {...props}
+  />
+)
+
+DoubleEntryTable.defaultProps = {
+  firstColWidth: 240,
+}
+
+DoubleEntryTable.propTypes = {
+  firstColWidth: PropTypes.number,
+}