+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add a11y events to `DropdownButton` component.
 - Fix: Fix `Dropdown` components events and accessibility.
 
+## [2.60.2] - 2020-03-26
+
+Fix:
+- More performant `DoubleEntryTable` Component.
+
 ## [2.60.1] - 2020-03-25
 
 Fix: